Only support Rails 6.1 now
Rails 6.0, 5.2, 5.1, 5.0 won't work with this version of ARS. In order
to get 6.1 compatibility working, we need remove distractions.
