+++ .github/workflows/ci.yml
@@ -27,28 +27,6 @@ jobs:
           - "2.6"
           - "2.7"
           - "3.0"
-        gemfile:
-          - rails5.0
-          - rails5.1
-          - rails5.2
-          - rails6.0
-        exclude:
-          - ruby-version: "2.7"
-            gemfile: rails5.0
-          - ruby-version: "2.7"
-            gemfile: rails5.1
-          - ruby-version: "2.7"
-            gemfile: rails5.2
-          - ruby-version: "3.0"
-            gemfile: rails4.2
-          - ruby-version: "3.0"
-            gemfile: rails5.0
-          - ruby-version: "3.0"
-            gemfile: rails5.1
-          - ruby-version: "3.0"
-            gemfile: rails5.2
-    env:
-      BUNDLE_GEMFILE: gemfiles/${{ matrix.gemfile }}.gemfile
     steps:
       - uses: zendesk/checkout@v2
       - name: Set up Ruby
