+++ .rubocop.yml
@@ -19,12 +19,6 @@ Layout/HashAlignment:
   EnforcedHashRocketStyle:
     - table
 
-Naming/FileName:
-  Exclude:
-    - 'lib/active_record_shards/connection_switcher-5-0.rb'
-    - 'lib/active_record_shards/connection_switcher-5-1.rb'
-    - 'lib/active_record_shards/connection_switcher-6-0.rb'
-
 Style/Alias:
   EnforcedStyle: prefer_alias_method
