+++ Gemfile
@@ -1,1 +1,18 @@
-eval_gemfile 'gemfiles/rails6.0.gemfile'
+# frozen_string_literal: true
+
+source 'https://rubygems.org'
+
+gemspec path: '.'
+
+gem 'bump'
+gem 'minitest', '>= 5.10.0'
+gem 'minitest-fail-fast'
+gem 'minitest-line'
+gem 'minitest-rg'
+gem 'mocha', '>= 1.4.0'
+gem 'mysql2'
+gem 'pry-byebug', platforms: [:mri]
+gem 'rake', '~> 12.0'
+gem 'rubocop', '~> 0.77.0'
+gem 'rubocop-minitest', '~> 0.5.0'
+gem 'rubocop-performance', '~> 1.5.1'
