--- gemfiles/rails5.0.gemfile
@@ -1,4 +1,0 @@
-eval_gemfile 'common.rb'
-
-gem 'activerecord', '~> 5.0.0'
-gem 'mysql2', '>= 0.3.18', '< 0.5'
