--- gemfiles/rails5.1.gemfile
@@ -1,4 +1,0 @@
-eval_gemfile 'common.rb'
-
-gem 'activerecord', '~> 5.1.3'
-gem 'mysql2', '>= 0.3.18', '< 0.5'
