--- gemfiles/rails6.0.gemfile
@@ -1,3 +1,0 @@
-eval_gemfile 'common.rb'
-
-gem 'activerecord', '~> 6.0.0'
